api/requirements.txt
# Kin Spark — AI Rep Product
# Python 3.12+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase
supabase>=2.3.0

# LLM
litellm>=1.30.0
google-generativeai>=0.4.0
anthropic>=0.18.0

# Embeddings (OpenAI text-embedding-3-large, 2000 dims)
openai>=1.12.0

# Async HTTP
httpx>=0.26.0

# JWT (admin portal auth — RS256 via JWKS)
PyJWT[crypto]>=2.8.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Dev Tools (Static Analysis)
black>=24.0.0
mypy>=1.8.0
bandit>=1.7.0
